extracted loading manager logic, dom element logic from 3d/world/index
