+++ client/3d/world/configure.js
