+++ client/3d/world/domElements.js
@@ -1,0 +1,7 @@
+
+export let HUD, LOADING_SCREEN
+
+export default () => {
+  HUD = document.getElementById('hudContainer')
+  LOADING_SCREEN = document.getElementById('progress-container')
+}
