+++ client/3d/world/index.js
@@ -1,55 +1,22 @@
 import store, {addPoints} from '../../store'
-import {configureRenderer} from '../configure'
+import {configureRenderer} from './configure'
 import {showInstructions} from '../../utilities'
-
-/*********************************
- * Construct the Three World
- ********************************/
+import getDomElements from './domElements'
+import loadingManager, {RESOURCES_LOADED} from './loadingManager'
 
 let isPaused = false
 let onEsc
-let loadingManager = null
-let RESOURCES_LOADED = false
-let counter = 0
 
-// An object to hold all the things needed for our loading screen
-var loadingScreen = {
-  scene: new THREE.Scene(),
-  camera: new THREE.PerspectiveCamera(90, 1280 / 720, 0.1, 100),
-  box: new THREE.Mesh(
-    new THREE.BoxGeometry(0.5, 0.5, 0.5),
-    new THREE.MeshBasicMaterial({color: 0x4444ff})
-  )
-}
-
-export default function generateWorld(/*world, currentUser, guestAvatar*/) {
+export default function generateWorld() {
   const {renderer, camera, scene, disposeOfResize} = configureRenderer()
+  getDomElements()
 
-  loadingManager = new THREE.LoadingManager()
-
-  var progress = document.getElementById('progress-container')
-  // progress.id = 'progress-container'
-  // var progressBar = document.getElementById('progress');
-  var HUD = document.getElementById('hudContainer')
 
   // progressBar.id = 'progress'
   // progress.appendChild(progressBar);
   // document.body.appendChild(progress);
 
-  let resources = 0
-  loadingManager.onProgress = function(item, loaded, total) {
-    // progressBar.style.width = (loaded / total * 100) + '%';
-    // console.log(`loaded resource ${loaded}/${total}`)
-    resources++
-  }
-
-  loadingManager.onLoad = function() {
-    console.log(`loaded ${resources} resources`)
-    RESOURCES_LOADED = true
-    // progressBar.style.display = 'none'
-    progress.style.display = 'none'
-    HUD.style.display = 'flex'
-  }
+  
 
   // loadingScreen.box.position.set(0,0,5);
   // loadingScreen.camera.lookAt(loadingScreen.box.position);
